Fixing the issue of the CI not responding
With Mark Elvers we identify where the issue about the CI not
responding when there's a lot of activity, meaning when there's a new
push in opam-repo the analysis of all user repositroy is restarted.

An analysis makes a tmp copy of an user repository in order to extract
all the information needed for the solver-service. There's some IO access
involved and take time when there's a lot of analysis running at same
time in the CI.

This is an attempt to fix that, by cache the extraction without keep
doing it for each new opam repository commit.